[api] Minor improvements
- Moved url function to api dir
- Added response status code check in tests
